+++ CmsCore/CmsCore.Model/Entities/PostCategory.cs
@@ -7,10 +7,10 @@ namespace CmsCore.Model.Entities
 {
     public class PostCategory
     {
-        public int PostId { get; set; }
+        public long PostId { get; set; }
         public virtual Post Post { get; set; }
 
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
         public virtual Category Category { get; set; }
     }
 }
